+++ src/jquery.fancytree.filter.js
@@ -89,6 +89,7 @@ $.ui.fancytree._FancytreeClass.prototype._applyFilterImpl = function(filter, bra
 	} else {
 		this.$div.addClass("fancytree-ext-filter-dimm");
 	}
+	this.$div.toggleClass("fancytree-ext-filter-hide-expanders", !!filterOpts.hideExpanders);
 	// Reset current filter
 	this.visit(function(node){
 		delete node.match;
@@ -281,14 +282,13 @@ $.ui.fancytree.registerExtension({
 	// Default options for this extension.
 	options: {
 		autoApply: true,  // Re-apply last filter if lazy data is loaded
-		counter: true,  // Show a badge with number of matching child nodes near parent icons
-		fuzzy: false,  // Match single characters in order, e.g. 'fb' will match 'FooBar'
-		hideExpandedCounter: true,  // Hide counter badge, when parent is expanded
+		counter: true,    // Show a badge with number of matching child nodes near parent icons
+		fuzzy: false,     // Match single characters in order, e.g. 'fb' will match 'FooBar'
+		hideExpandedCounter: true,  // Hide counter badge if parent is expanded
+		hideExpanders: true,        // Hide expanders if all child nodes are hidden by filter
 		highlight: true,  // Highlight matches by wrapping inside <mark> tags
-		nodata: true,  // Display a 'no data' status node if result is empty
-		mode: "dimm",  // Grayout unmatched nodes (pass "hide" to remove unmatched node instead)
-		// events
-		enhanceTitle: null
+		nodata: true,     // Display a 'no data' status node if result is empty
+		mode: "dimm"      // Grayout unmatched nodes (pass "hide" to remove unmatched node instead)
 	},
 	nodeLoadChildren: function(ctx, source) {
 		return this._superApply(arguments).done(function() {