+++ CHANGELOG.md
@@ -7,13 +7,14 @@
   * [Added] #620 improved tooltip escaping to allow newlines
   * [DEPRECATED] `removeNode` event. Listen for `modifyChild` with operation
     'remove' instead (which is fired on the parent).
-  * [Improved] #230 themeroller theme compatible with ext-filter
+  * [Improved] ThemeRoller theme
   * [Improved] #656 WAI-ARIA support
     - Set focus to first node on first tab-in
     - Support [home] and [end] keys
     - Set aria-activedescendant on container to active ID
     - Set aria-multiselectable on container if selectMode != 1
     - Set aria-treeitem, -selected, -expanded, on title span instead `<li>`
+  * [Fixed] #230 themeroller theme compatible with ext-filter
   * [Fixed] #528 autoCollapse option blocks filter's autoExpand option
   * [Fixed] #576 `loadKeyPath()` sometimes gets the root wrong
   * [Fixed] #615 Drag & drop helper icons lose indentation with table extension